fix: fixed prop name on nftunes, moved ytembed component into components folder, changed the fullscreen attributes
